+++ src/drawable.rs
@@ -1,7 +1,7 @@
 use crate::color::Color;
 use crate::shapes::CreateDrawable;
 use graphics_shapes::coord::Coord;
-use graphics_shapes::Shape;
+use graphics_shapes::{IntersectsContains, Shape};
 #[cfg(feature = "serde_derive")]
 use serde::{Deserialize, Serialize};
 
@@ -82,12 +82,17 @@ where
     T: Shape + Clone,
 {
     pub fn with_translation<P: Into<Coord>>(&self, delta: P) -> Drawable<T> {
-        let moved = self.obj.translate_by(delta);
+        let moved = self.obj.translate_by(delta.into());
         Drawable::from_obj(moved, self.draw_type)
     }
 
     pub fn with_move<P: Into<Coord>>(&self, xy: P) -> Drawable<T> {
-        let moved = self.obj.move_to(xy);
+        let moved = self.obj.move_to(xy.into());
+        Drawable::from_obj(moved, self.draw_type)
+    }
+
+    pub fn with_move_center<P: Into<Coord>>(&self, xy: P) -> Drawable<T> {
+        let moved = self.obj.move_center_to(xy.into());
         Drawable::from_obj(moved, self.draw_type)
     }
 
@@ -97,7 +102,7 @@ where
     }
 
     pub fn with_scale_around<P: Into<Coord>>(&self, scale: f32, point: P) -> Drawable<T> {
-        let moved = self.obj.scale_around(scale, point);
+        let moved = self.obj.scale_around(scale, point.into());
         Drawable::from_obj(moved, self.draw_type)
     }
 
@@ -107,7 +112,45 @@ where
     }
 
     pub fn with_rotation_around<P: Into<Coord>>(&self, degrees: isize, point: P) -> Drawable<T> {
-        let rotated = self.obj.rotate_around(degrees, point);
+        let rotated = self.obj.rotate_around(degrees, point.into());
         Drawable::from_obj(rotated, self.draw_type)
+    }
+
+    pub fn left(&self) -> isize {
+        self.obj.left()
+    }
+
+    pub fn right(&self) -> isize {
+        self.obj.right()
+    }
+
+    pub fn top(&self) -> isize {
+        self.obj.top()
+    }
+
+    pub fn bottom(&self) -> isize {
+        self.obj.bottom()
+    }
+}
+
+impl<T> Drawable<T>
+where
+    Self: CreateDrawable<T>,
+    T: IntersectsContains + Clone,
+{
+    pub fn intersects_shape(&self, shape: &dyn Shape) -> Option<bool> {
+        self.obj.intersects_shape(shape)
+    }
+
+    pub fn contains_shape(&self, shape: &dyn Shape) -> Option<bool> {
+        self.obj.contains_shape(shape)
+    }
+
+    pub fn intersects_drawable<S: Shape + Clone>(&self, drawable: &Drawable<S>) -> Option<bool> {
+        self.obj.intersects_shape(&drawable.obj)
+    }
+
+    pub fn contains_drawable<S: Shape + Clone>(&self, drawable: &Drawable<S>) -> Option<bool> {
+        self.obj.contains_shape(&drawable.obj)
     }
 }